2005-05-02	Jennifer Averett <jennifer.averett@oarcorp.com>
	* libchip/network/cs8900.c, libchip/network/i82586.c,
	libchip/network/open_eth.c, libchip/network/sonic.c,
	libchip/serial/mc68681.c: Modified types to remove
	warnings.  Added initialization to remove warnings.

+++ c/src/libchip/network/cs8900.c
@@ -1079,7 +1079,7 @@ cs8900_init (void *arg)
 }
 
 static int
-cs8900_ioctl (struct ifnet *ifp, int cmd, caddr_t data)
+cs8900_ioctl (struct ifnet *ifp, u_long cmd, caddr_t data)
 {
   cs8900_device *cs = ifp->if_softc;
   int           error = 0;
